delete task_2_3_1 into package path
