+++ Task_2_3_1/build.gradle
@@ -22,8 +22,8 @@ tasks.withType(JavaCompile) {
 }
 
 application {
-    mainClass = 'ru.nsu.seleznev.a.task_2_3_1.SnakeGame'
-    mainModule = 'ru.nsu.seleznev.a.task_2_3_1'
+    mainClass = 'ru.nsu.seleznev.a.SnakeGame'
+    mainModule = 'ru.nsu.seleznev.a'
 }
 
 javafx {
