+++ Task_2_3_1/src/main/java/ru/nsu/seleznev/a/view/Background.java
@@ -1,4 +1,4 @@
-package ru.nsu.seleznev.a.task_2_3_1.view;
+package ru.nsu.seleznev.a.view;
 
 
 import javafx.scene.canvas.GraphicsContext;
