+++ Task_2_3_1/src/test/java/ru/nsu/seleznev/a/EmptyTest.java
@@ -1,9 +1,12 @@
-package ru.nsu.seleznev.a.task_2_3_1;
+package ru.nsu.seleznev.a;
 
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+/**
+ * Empty test for fun.
+ */
 public class EmptyTest {
   @Test
   public void emptyTest() {
